Replace pw_trigger with simple_trigger.
Former is meant for PW-USB. Here, the normal destination for the trace
trigger is Husky, which will have its own programmable trigger
infrastructure, so no need to duplicate it here.
Also, pw_trigger was written for the PW-USB clocking scenario where
trigger_clk = 4x, which is not the case for trace.